- uses Framework for updateCard - update Categories doesn't work anymore - included bootstrap

+++ src/main/java/de/thb/anylearn/repository/FolderRepository.java
@@ -4,7 +4,7 @@ import de.thb.anylearn.entity.Folder;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.RepositoryDefinition;
 
-@RepositoryDefinition(domainClass = Folder.class, idClass = Long.class)
-public interface FolderRepository extends CrudRepository<Folder, Long> {
+@RepositoryDefinition(domainClass = Folder.class, idClass = Integer.class)
+public interface FolderRepository extends CrudRepository<Folder, Integer> {
 
 }
